server/requirements.txt
asgiref==3.2.3
Django==3.0.7
django-cors-headers===3.3.0
djangorestframework==3.11.0
git+git://github.com/AtuzSolution/django-rest-framework-simplejwt@jwt_6#egg=djangorestframework_simplejwt
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
